+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - add support for new 20.15 parameters in system OMP, Global and SNMP features
 - add support for transport ipsec interface
 - rename `ipv4_configuration_type`/`ipv6_configuration_type` to `ipv4_address_type`/`ipv6_address_type` and add `ipv4_address_type_variable`/`ipv6_address_type_variable` in service LAN VPN ethernet interface, transport management VPN ethernet interface and transport WAN VPN ethernet interface
+- add support for mrf, port-channel and other new 20.15 parameters for wan_vpn ethernet interface
 
 ## 1.3.0